@@ -1,24 +1,31 @@
 {
   "name": "apexcharts",
-  "description": "A JavaScript Chart Library",
   "version": "2.6.0",
-  "main": "dist/apexcharts.min.js",
-  "typings": "./apexcharts.d.ts",
-  "directories": {
-    "test": "tests"
+  "description": "A JavaScript Chart Library",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/apexcharts/apexcharts.js.git"
   },
+  "main": "dist/apexcharts.common.js",
+  "module": "dist/apexcharts.esm.js",
+  "unpkg": "dist/apexcharts.js",
+  "jsdelivr": "dist/apexcharts.js",
+  "typings": "types/apexcharts.d.ts",
+  "files": [
+    "src",
+    "dist/*.js",
+    "types/*.d.ts"
+  ],
   "scripts": {
-    "start": "webpack --watch",
-    "bundle": "webpack",
-    "rollup": "rollup -c ./rollup.config.js",
-    "minify": "gulp build",
-    "build": "npm run bundle && npm run rollup && npm run minify",
+    "dev": "rollup -w -c build/config.js --environment TARGET:web-umd-dev",
+    "dev:cjs": "rollup -w -c build/config.js --environment TARGET:web-cjs",
+    "dev:esm": "rollup -w -c build/config.js --environment TARGET:web-esm",
+    "build": "node build/build.js",
+    "build:umd": "rollup -w -c build/config.js --environment TARGET:web-umd-dev",
+    "lint": "eslint src/**/*.js tests/**/*.js --no-fix",
     "test": "jest"
   },
   "dependencies": {
-    "babel-polyfill": "^6.26.0",
-    "core-js": "^2.5.7",
-    "es6-promise-promise": "^1.0.0",
     "svg.draggable.js": "^2.2.1",
     "svg.easing.js": "^2.0.0",
     "svg.filter.js": "^2.0.2",
@@ -28,46 +35,34 @@
     "svg.select.js": "^2.1.2"
   },
   "devDependencies": {
-    "babel-core": "^6.9.0",
-    "babel-eslint": "^8.2.2",
-    "babel-jest": "^23.4.2",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-transform-class-properties": "^6.9.0",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "concurrently": "^3.5.1",
-    "css-loader": "^0.28.10",
-    "eslint": "^4.18.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-config-standard": "^11.0.0",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-import": "^2.9.0",
-    "eslint-plugin-node": "^6.0.1",
-    "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "exec-sh": "^0.3.2",
-    "gulp": "^3.9.1",
-    "gulp-remove-logging": "^1.2.0",
-    "gulp-rename": "^1.4.0",
-    "gulp-strip-debug": "^3.0.0",
-    "gulp-uglify": "^3.0.1",
-    "imports-loader": "^0.7.1",
-    "jest": "^23.5.0",
-    "jest-puppeteer": "^3.3.1",
-    "puppeteer": "^1.7.0",
-    "puppeteer-core": "^1.7.0",
-    "rollup": "^0.62.0",
-    "rollup-plugin-commonjs": "^9.1.3",
-    "rollup-plugin-node-resolve": "^3.4.0",
-    "rollup-plugin-uglify": "^6.0.0",
-    "rollup-plugin-uglify-es": "0.0.1",
-    "style-loader": "^0.20.2",
-    "svg-inline-loader": "^0.8.0",
-    "webpack": "^2.2.1"
+    "@babel/core": "7.2.2",
+    "@babel/plugin-proposal-class-properties": "7.2.3",
+    "@babel/preset-env": "7.2.3",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-eslint": "10.0.1",
+    "babel-jest": "23.6.0",
+    "chalk": "2.4.2",
+    "cross-env": "5.2.0",
+    "eslint": "5.12.0",
+    "eslint-config-prettier": "3.3.0",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-prettier": "3.0.1",
+    "eslint-plugin-promise": "4.0.1",
+    "jest": "23.6.0",
+    "jest-puppeteer": "3.8.0",
+    "prettier": "1.15.3",
+    "puppeteer": "1.11.0",
+    "puppeteer-core": "1.11.0",
+    "rollup": "1.1.0",
+    "rollup-plugin-babel": "4.3.0",
+    "rollup-plugin-json": "3.1.0",
+    "rollup-plugin-node-resolve": "4.0.0",
+    "rollup-plugin-postcss": "1.6.3",
+    "rollup-plugin-replace": "2.1.0",
+    "rollup-plugin-svgo": "1.0.2",
+    "terser": "3.14.1",
+    "tslint": "5.12.1",
+    "typescript": "3.2.2"
   },
   "author": {
     "name": "Juned Chhipa",
